Add custom commands for code lisibility
Add custom commands for mailhog interactions with Cypress
