+++ client/tests/e2e/specs/adminPlaces.js
@@ -1,4 +1,3 @@
-import mesCandi from '../../../src/candidat'
 /* Tests :
 - Display of colours when there is or not available places on the dashboard
   - Red if less than half of the places from the week have been reserved
@@ -23,75 +22,10 @@ describe('Places tests', () => {
   before(() => {
     // Delete all mails before start
     cy.mhDeleteAll()
-    // Creates the aurige files
-    cy.writeFile(Cypress.env('filePath') + '/aurige.json',
-      [
-        {
-          'codeNeph': Cypress.env('NEPH'),
-          'nomNaissance': Cypress.env('candidat'),
-          'email': Cypress.env('emailCandidat'),
-          'dateReussiteETG': '2018-10-12',
-          'nbEchecsPratiques': '0',
-          'dateDernierNonReussite': '',
-          'objetDernierNonReussite': '',
-          'reussitePratique': '',
-          'candidatExistant': 'OK',
-        },
-      ])
-    cy.writeFile(Cypress.env('filePath') + '/aurige.end.json',
-      [
-        {
-          'codeNeph': Cypress.env('NEPH'),
-          'nomNaissance': Cypress.env('candidat'),
-          'email': Cypress.env('emailCandidat'),
-          'dateReussiteETG': '',
-          'nbEchecsPratiques': '',
-          'dateDernierNonReussite': '',
-          'objetDernierNonReussite': '',
-          'reussitePratique': '',
-          'candidatExistant': 'NOK',
-        },
-      ])
+    cy.adminLogin()
+    cy.archiveCandidate()
     // Creates the csv file
     cy.writeFile(Cypress.env('filePath') + '/planning.csv', placesArray)
-    // Archives the candidate if it's not already done
-    cy.visit(Cypress.env('frontAdmin') + 'admin-login')
-    cy.get('[type=text]')
-      .type(Cypress.env('adminLogin'))
-    cy.get('[type=password]')
-      .type(Cypress.env('adminPass'))
-    cy.get('.submit-btn')
-      .click()
-    cy.get('.v-snack')
-      .should('contain', 'Vous êtes identifié')
-    cy.contains('import_export')
-      .click()
-    const filePath = '../../../' + Cypress.env('filePath') + '/aurige.end.json'
-    const fileName = 'aurige.json'
-    cy.fixture(filePath).then(fileContent => {
-      cy.get('.input-file-container [type=file]')
-        .upload({
-          fileContent: JSON.stringify(fileContent),
-          fileName,
-          mimeType: 'application/json',
-        })
-    })
-    cy.get('.v-snack')
-      .should('contain', fileName + ' prêt à être synchronisé')
-    cy.get('.import-file-action [type=button]')
-      .click()
-    cy.get('.v-snack')
-      .should('contain', 'Le fichier ' + fileName + ' a été synchronisé.')
-    // Adds the address to the whitelist
-    cy.visit(Cypress.env('frontAdmin') + 'admin/whitelist')
-    cy.get('h2')
-      .should('contain', 'Liste blanche')
-    cy.contains('Ajouter un lot d\'adresse courriel')
-      .click()
-    cy.get('#whitelist-batch-textarea')
-      .type(Cypress.env('emailCandidat'))
-    cy.contains('Enregistrer ces adresses')
-      .click()
     // Adds the places from the created planning file
     cy.contains('calendar_today')
       .click()
@@ -108,138 +42,18 @@ describe('Places tests', () => {
       .click({ force: true })
     cy.get('.v-snack', { timeout: 10000 })
       .should('contain', 'Le fichier ' + fileName1 + ' a été traité pour le departement 75.')
-    cy.get('.t-disconnect')
-      .click()
-    cy.get('.v-snack')
-      .should('contain', 'Vous êtes déconnecté·e')
-    // The candidate fills the pre-sign-up form
-    cy.visit(Cypress.env('frontCandidat') + 'qu-est-ce-que-candilib')
-    cy.contains('Se pré-inscrire')
-      .click()
-    cy.get('h2')
-      .should('contain', mesCandi.app_subtitle)
-    cy.contains(mesCandi.preinscription_neph)
-      .parent()
-      .children('input')
-      .type(Cypress.env('NEPH'))
-    cy.contains(mesCandi.preinscription_nom_naissance)
-      .parent()
-      .children('input')
-      .type(Cypress.env('candidat'))
-    cy.contains(mesCandi.preinscription_prenom)
-      .parent()
-      .children('input')
-      .type(Cypress.env('firstName'))
-    cy.contains('Courriel *')
-      .parent()
-      .children('input')
-      .type(Cypress.env('emailCandidat'))
-    cy.contains(mesCandi.preinscription_mobile)
-      .parent()
-      .children('input')
-      .type('0716253443')
-    cy.contains(mesCandi.preinscription_adresse)
-      .parent()
-      .children('input')
-      .type('avenue')
-    cy.get('.v-select-list')
-      .contains('avenue')
-      .click()
-    cy.contains(mesCandi.preinscription_bouton_submit)
-      .click()
-    // Checks the access
-    cy.url()
-      .should('contain', 'email-validation')
-    cy.get('h3')
-      .should('contain', 'Validation en attente')
-    cy.get('div')
-      .should('contain', 'Vous allez bientôt recevoir un courriel à l\'adresse que vous nous avez indiqué.')
-    // Validates the email address
-    cy.mhGetAllMails()
-      .mhFirst()
-      .mhGetRecipients()
-      .should('contain', Cypress.env('emailCandidat'))
-    cy.mhGetAllMails()
-      .mhFirst()
-      .mhGetSubject()
-      .should('contain', 'Validation d\'adresse courriel pour Candilib')
-    cy.mhGetAllMails()
-      .mhFirst()
-      .mhGetBody().then((mailBody) => {
-        const codedLink = mailBody.split('href=3D"')[1].split('">')[0]
-        const withoutEq = codedLink.replace(/=\r\n/g, '')
-        const validationLink = withoutEq.replace(/=3D/g, '=')
-        cy.visit(validationLink)
-      })
-    cy.get('h3')
-      .should('contain', 'Adresse courriel validée')
-    // Gets the confirmation email
-    cy.mhGetAllMails()
-      .mhFirst()
-      .mhGetRecipients()
-      .should('contain', Cypress.env('emailCandidat'))
-    cy.mhGetAllMails()
-      .mhFirst()
-      .mhGetSubject()
-      .should('contain', '=?UTF-8?Q?Inscription_Candilib_en_attente_de_v?= =?UTF-8?Q?=C3=A9rification?=')
+    cy.addToWhitelist()
+    cy.adminDisconnection()
+    cy.candidatePreSignUp()
     // The admin validates the candidate via Aurige
-    cy.visit(Cypress.env('frontAdmin') + 'admin-login')
-    cy.get('[type=text]')
-      .type(Cypress.env('adminLogin'))
-    cy.get('[type=password]')
-      .type(Cypress.env('adminPass'))
-    cy.get('.submit-btn')
-      .click()
-    cy.get('.v-snack')
-      .should('contain', 'Vous êtes identifié')
-    cy.contains('import_export')
-      .click()
-    cy.get('.ag-overlay')
-      .should('contain', 'No Rows To Show')
-    const filePath2 = '../../../' + Cypress.env('filePath') + '/aurige.json'
-    const fileName2 = 'aurige.json'
-    cy.fixture(filePath2).then(fileContent => {
-      cy.get('.input-file-container [type=file]')
-        .upload({
-          fileContent: JSON.stringify(fileContent),
-          fileName: fileName2,
-          mimeType: 'application/json',
-        })
-    })
-    cy.get('.v-snack')
-      .should('contain', fileName2 + ' prêt à être synchronisé')
-    cy.get('.import-file-action [type=button]')
-      .click()
-    cy.get('.v-snack')
-      .should('contain', 'Le fichier ' + fileName2 + ' a été synchronisé.')
-    // Checks that the candidate is validated
-    cy.get('.ag-cell')
-      .should('contain', Cypress.env('candidat'))
-    cy.get('.ag-cell')
-      .should('contain', 'Pour le 75, envoi d\'un magic link à ' + Cypress.env('emailCandidat'))
-    cy.mhGetMailsByRecipient(Cypress.env('emailCandidat'))
-      .mhFirst()
-      .mhGetSubject()
-      .should('contain', '=?UTF-8?Q?Validation_de_votre_inscription_=C3=A0_C?= =?UTF-8?Q?andilib?=')
+    cy.adminLogin()
+    cy.candidateValidation()
     // Disconnects from the app
-    cy.get('.t-disconnect')
-      .click()
-    cy.get('.v-snack')
-      .should('contain', 'Vous êtes déconnecté·e')
-    cy.url()
-      .should('eq', Cypress.env('frontAdmin') + 'admin-login')
+    cy.adminDisconnection()
   })
 
   it('Checks the colours on the dashboard', () => {
-    cy.visit(Cypress.env('frontAdmin') + 'admin-login')
-    cy.get('[type=text]')
-      .type(Cypress.env('adminLogin'))
-    cy.get('[type=password]')
-      .type(Cypress.env('adminPass'))
-    cy.get('.submit-btn')
-      .click()
-    cy.get('.v-snack')
-      .should('contain', 'Vous êtes identifié')
+    cy.adminLogin()
     // Checks the places on the dashboard
     cy.get('h2.title')
       .should('contain', Cypress.env('centre'))
