+++ client/tests/e2e/support/mailHogCommands.js
@@ -1,0 +1,17 @@
+Cypress.Commands.add('mhGetFirstRecipients', () => {
+  cy.mhGetAllMails()
+    .mhFirst()
+    .mhGetRecipients()
+})
+
+Cypress.Commands.add('mhGetFirstSubject', () => {
+  cy.mhGetAllMails()
+    .mhFirst()
+    .mhGetSubject()
+})
+
+Cypress.Commands.add('mhGetFirstBody', () => {
+  cy.mhGetAllMails()
+    .mhFirst()
+    .mhGetBody()
+})
